testing the bbox input params for query carefully
Splitting the >1000 lines test modules into separate files
